API 5L. Values of SMYS and UTS

Data represents specified minimum yield strength (SMYS) and ultimate tensile strength (UTS) both achieved when producing line pipes according to regulatory specifications.

Usage

api5l3tdata

Format

A data frame with 57 rows and 4 variables:

grade

designation of standard grade of manufactured pipe. Type: assert_character.

smys

SMYS - specified minimum yield strength, [psi]. Type: assert_double.

uts

UTS - ultimate tensile strength, [psi]. Type: assert_double.

origin

Identifier for the information origin regarding the specifications of pipe, []:

10

API SPECIFICATION 5L .Table 3A

3

GOST 20295-85. Table 7

11

GOST 31443-2012 . Tables 6, 7

Type: assert_integer.

References

API SPECIFICATION 5L. Specification for Line Pipe

GOST 20295-85. Steel welded pipes for main gas-and-oil pipelines. Specifications

GOST 31443-2012. Steel pipes for crafts pipelines. Specifications.

Examples

library(pipenostics)

head(api5l3tdata)
#>   grade  smys   uts origin
#> 1   A25 25000 45000     10
#> 2     A 30000 48000     10
#> 3     B 35000 60000     10
#> 4   X42 42000 60000     10
#> 5   X46 46000 63000     10
#> 6   X52 52000 66000     10