ASME B31G. Allowable corrosion length in pipe

Calculate allowable length of the corroded area in the pipe.

b31gacl(dep, maop, d, wth, depth, l)

Arguments

dep
design pressure of the pipe, [PSI]. Type: assert_double.

maop
maximum allowable operating pressure - MAOP, [PSI]. Type: assert_double.

d
nominal outside diameter of the pipe, [inch]. Type: assert_double.

wth
nominal wall thickness of the pipe, [inch]. Type: assert_double.

depth
measured maximum depth of the corroded area, [inch]. Type: assert_double.

l
measured maximum longitudinal length of the corroded area, [inch]. Type: assert_double.

Value

allowable depth of the corroded area in the pipe, [inch]. Type: assert_double.

References

ASME B31G-1991. Manual for determining the remaining strength of corroded pipelines. A supplement to ASTME B31 code for pressure piping.

See also

Examples

 b31gacl(1093, 910, 30, .438, .1, 7.5)

#> [1] Inf

 # [1] Inf  # [inch] - corrosion is low, no limit for the corroded area length

 b31gacl(438, 400, 20, .25, .18, 10)

#> [1] 2.018

 # [1] 2.018  # [inch] - finite allowed length of the corroded area

 ## unit test:
 data(b31gdata)
 with(b31gdata,
      stopifnot(all.equal(
        b31gacl(design_pressure, maop, d, wth, depth, l),
        allowed_corrosion_length)))