ASME B31G. A-factor

Calculate intermediate factor related to the geometry of the corroded zone.

b31gafr(d, wth, l)

Arguments

d
nominal outside diameter of the pipe, [inch]. Type: [double].

wth
nominal wall thickness of the pipe, [inch]. Type: [double].

l
measured maximum longitudial length of the corroded area, [inch]. Type: [double].

Value

Intermediate factor related to the geometry of the corroded area, []. Type: [double].

References

ASME B31G-1991. Manual for determining the remaining strength of corroded pipelines. A supplement to ASTME B31 code for pressure piping.

See also

Examples

 b31gafr(30, .438, 7.5)

#> [1] 1.847

 # [1] 1.847  # A-factor is less than 5, so the corrosion is not critical

 ## unit test:
 data(b31gdata)
 with(b31gdata, all.equal(b31gafr(d, wth, l), A))

#> [1] TRUE