ASME B31G. Corrosion state of 12 pipes

Data represents examples used for verification of computer program CRVL.BAS listed in Appendix A of ASME B31G-1991.

Usage

b31gdata

Format

A data frame with 12 rows and 15 variables:

maop

maximum allowable operating pressure - MAOP, [PSI]. Type: assert_double.

d

nominal outside diameter of pipe, [inch]. Type: assert_double.

wth

nominal wall thickness of pipe, [inch]. Type: assert_double.

smys

specified minimum yield of stress (SMYS) as a characteristics of steel strength, [PSI]. Type: assert_double.

def

appropriate (combined) design factor from ASME B31.4, ASME B31.8, or ASME B31.11, []. Type: assert_double.

depth

measured maximum depth of the corroded area, [inch]. Type: assert_double.

l

measured maximum longitudinal length of corroded area, [inch]. Type: assert_double.

status

operational status of pipe: 1 - excellent, 2 - monitoring is recommended, 3 - alert! replace the pipe immediately! Type: assert_numeric.

design_pressure

design pressure of pipe, [PSI]. Type: assert_double.

safe_pressure

safe maximum pressure for the corroded area, [PSI]. Type: assert_double.

pressure_exceeding

whether operator's action is required to reduce MOAP lower than the maximum safe pressure of the corroded area. Type: assert_logical.

allowed_corrosion_depth

allowable depth of the corroded area, [inch]. Type: assert_double.

A

intermediate factor related to the geometry of the corroded area, []. Type: assert_double.

allowed_corrosion_length

allowable length of the corroded area, [inch]. Type: assert_double.

AP

another intermediate factor related to the geometry of the corroded area, []. Type: assert_double.

Source

https://law.resource.org/pub/us/cfr/ibr/002/asme.b31g.1991.pdf

Examples

library(pipenostics)

head(b31gdata)
#>           maop  d   wth  smys  def depth     l status design_pressure
#> Example 1  910 30 0.438 52000 0.72  0.10   7.5      2            1093
#> Example 2  400 20 0.250 35000 0.50  0.18  10.0      2             438
#> Example 3  910 24 0.432 52000 0.72  0.13  30.0      2            1348
#> Example 4  910 24 0.432 52000 0.72  0.30  30.0      2            1348
#> Example 5  731 24 0.281 52000 0.72  0.08  15.0      2             877
#> Example 6 1000 36 0.500 52000 0.72  0.41 100.0      3            1040
#>           safe_pressure pressure_exceeding allowed_corrosion_depth      A
#> Example 1          1093              FALSE                   0.249  1.847
#> Example 2           284               TRUE                   0.079  3.993
#> Example 3          1037              FALSE                   0.167  8.320
#> Example 4           453               TRUE                   0.167  8.320
#> Example 5           690               TRUE                   0.068  5.158
#> Example 6           206               TRUE                   0.063 21.048
#>           allowed_corrosion_length    AP
#> Example 1                      Inf 5.000
#> Example 2                    2.018 0.806
#> Example 3                      Inf 5.000
#> Example 4                   12.867 3.568
#> Example 5                   11.634 4.000
#> Example 6                    2.556 0.538