ASME B31G. Design pressure of pipe

Calculate the design pressure that according to ASME B31G-1991 is the conditioned construction characteristic that should not in no way exceeded.

Usage

b31gdep(d, wth, smys, def)

Arguments

d

nominal outside diameter of pipe, [inch]. Type: assert_double.

wth

nominal wall thickness of pipe, [inch]. Type: assert_double.

smys

specified minimum yield of stress (SMYS) as a characteristics of steel strength, [PSI]. Type: assert_double.

def

appropriate (combined) design factor from ASME B31.4, ASME B31.8, or ASME B31.11, []. Type: assert_double.

Value

Design pressure of pipe, [PSI]. Type: assert_double.

References

ASME B31G-1991. Manual for determining the remaining strength of corroded pipelines. A supplement to ASTME B31 code for pressure piping.

See also

Examples

library(pipenostics)

b31gdep(30, .438, 52e3, .72)
#> [1] 1093.748