ASME B31G. Safe maximum pressure for the corroded area of pipe

Calculate safe maximum pressure for the corroded area of pipe.

b31gsap(dep, d, wth, depth, l)

Arguments

dep

design pressure of the pipe, [PSI]. Type: assert_double.

d

nominal outside diameter of the pipe, [inch]. Type: assert_double.

wth

nominal wall thickness of the pipe, [inch]. Type: assert_double.

depth

measured maximum depth of the corroded area, [inch]. Type: assert_double.

l

measured maximum longitudinal length of the corroded area, [inch]. Type: assert_double.

Value

Safe maximum pressure for the corroded area of pipe, [PSI]. Type: assert_double.

References

ASME B31G-1991. Manual for determining the remaining strength of corroded pipelines. A supplement to ASTME B31 code for pressure piping.

See also

Examples

 b31gsap(1093, 30, .438, .1, 7.5)
#> [1] 1093
 # [1] 1093  # [PSI], safe pressure is equal to design pressure

 b31gsap(877, 24, .281, .08, 15)
#> [1] 690
 # [1] 690   # [PSI], safe pressure is lower than design pressure due corrosion