ASME B36.10M. Estimate pipe mass

Estimate mass or geometric specifications of the manufactured pipes

Usage

b36mass(d, wth, len = 1, rho = 7.85, origin = NULL)

b36d(wth, mass, len = 1, rho = 7.85)

b36wth(d, mass, len = 1, rho = 7.85)

Arguments

d

nominal outside diameter of pipe, [mm]. Type: assert_double

wth

nominal wall thickness of pipe, [mm]. Type: assert_double.

len

pipe length, [m]. Type: assert_double.

rho

mass density of pipe material, [g/cm³]. Type: assert_double.

origin

identifier for the information origin regarding the specifications of pipe. Type: assert_integer.

mass

actual mass of pipe, [kg]. Type: assert_double.

Value

for b36mass

pipe mass, [kg]

for b36wth

pipe wall thickness, [mm]

for b36d

outside diameter of pipe, [mm]

Details

The mass of a one-meter pipe segment is determined by linear interpolation based on the tabular data provided in the specified origins (see b36pipedata). If the origin of the initial data is NULL (default), the mass of the pipe is calculated using a formula:

$$M = 10^{-3} \pi \rho w \left(d - w \right ) \cdot l$$

where

\(M\) - mass of a pipe, [kg]

\(\rho\) - mass density of pipe material, [g/cm³]

\(w\) - nominal wall thickness of the manufactured pipe, [mm]

\(d\) - nominal outside diameter of the manufactured pipe, [mm]

\(l\) - actual pipe length, [m]

For origin in c(1, 2, 4, 5) the values provided in the b36pipedata match the calculated values obtained from the formula with an accuracy of 1 %.

Inverse calculations b36wth and b36d are performed using algebraically derived formulas.

See also

Examples

library(pipenostics)

# Since some specification origins provide the mass of a one-meter pipe
# segment taking into account possible deviations during its production
# process, when the user specifies the origin ID directly, the mass values
# are calculated using linear interpolation:
b36mass(68, 13, rho = 7.9, origin = 7L)
#> [1] 16.43

# The discrepancy with the calculations based on the formula can be more
# than 7 %:
b36mass(68, 13, rho = 7.9, origin = NULL)
#> [1] 17.74529

# For origins which are ASME B36 standards such differences should be
# minimal:
b36mass(965, 10.31, origin = 1L) - b36mass(965, 10.31, origin = NULL)
#> [1] 0.0004356046

# The calculations of diameter and wall thickness are straightforward and
# use only inverse formulas without origin references:
b36d(10.31, 242.74)
#> [1] 965.0017

b36wth(965, 242.74)
#> [1] 10.31002