ASME B36.10M. Specifications of the manufactured pipes

Data represents the nominal specifications of steel pipes produced by the industry according to regulatory standards.

Usage

b36pipedata

Format

A data frame with 6064 rows and 5 variables:

d

Nominal outside diameter of the manufactured pipe, [mm]. Type: assert_double.

wth

Nominal wall thickness of the manufactured pipe, [mm]. Type: assert_double.

rho

Nominal mass density of the steel rank applied in the pipe manufacturing, [g/cm³]. Type: assert_double.

mass

Nominal mass of one-meter length pipe segment, [kg]. Type: assert_double.

origin

1

ASME B36.10M-2018

2

ASME B36.19M-2018

3

GOST 20295-85. Table 1

4

GOST 33229-2015. Table 1

5

GOST 33229-2015. Table 2

6

GOST R 57423-2017. Table 1

7

GOST R 57423-2017. Table 2

8

GOST R 57423-2017. Table 3

9

GOST R 57423-2017. Table 4

Type: assert_integer.

NOTE! Due to numerous typos in origins with identifiers 4, 5 all mass values in those origins are the recalculations made with formula

$$M = 0.02466 \cdot w\left(d - w \right )$$

where

\(M\) - mass of one-meter pipe segment, [kg]

\(d\) - nominal outside diameter of the manufactured pipe, [mm]

\(w\) (wth) - nominal wall thickness of the manufactured pipe, [mm]

References

ASME B36.10M-2018. Welded and seamless wrought steel pipe.

ASME B36.19M-2018. Stainless steel pipe.

GOST 20295-85. Steel welded pipes for main gas-and-oil pipelines. Specifications.

GOST 33229-2015. Tubes for boiler and heat exchanging equipment. Technical specifications. Part 1. Seamless steel pipes to work under pressure. not more than 6,4 MPa and at temperatures not exceeding 400 °C

GOST R 57423-2017. Tubes for boiler and heat exchanging equipment. Part 2. Seamless steel tubes for pressure purposes more 6.4 MPa and temperatures exceeding 400 °C. Specifications.

See also

Examples

library(pipenostics)

head(b36pipedata)
#>      d  wth  rho mass origin
#> 1 10.3 1.24 7.85 0.28      1
#> 2 10.3 1.45 7.85 0.32      1
#> 3 10.3 1.73 7.85 0.37      1
#> 4 10.3 2.41 7.85 0.47      1
#> 5 10.3 3.15 7.85 0.56      1
#> 6 10.3 4.83 7.85 0.65      1