Estimate pipe friction factor with Vatankhah formula

Estimate Darcy friction factor explicitly with extremely accurate Vatankhah-Kouchakzadeh approximation of Colebrook equation.

fric_vatankhan(reynolds, roughness = 0, strict = FALSE)

Arguments

reynolds

Reynolds number, []. Type: assert_double.

roughness

relative roughness, []. Type: assert_double.

strict

calculate only inside the precision region. Type: assert_flag.

Value

pipe friction factor, []. Type: assert_double.

Details

Vatankhah's formula is reported to be extremely accurate in the region:

5.0e3 <= reynolds <= 1.0e8

1e-6 <= roughness <= 0.05

In strict = TRUE mode argument values outside this precision region are not allowed, whereas in strict = FALSE either NAs are generated in that case or calculation for laminar flow is performed when reynolds < 2100.0.

References

Offor, U. and Alabi, S. (2016) An Accurate and Computationally Efficient Explicit Friction Factor Model. Advances in Chemical Engineering and Science, 6, pp. 237-245. DOI: 10.4236/aces.2016.63024.

Ali R. Vatankhah, Salah Kouchakzadeh (2009) Exact equations for pipe-flow problems. Journal of Hydraulic Research, 47:4, pp. 537-538, DOI: 10.1080/00221686.2009.9522031

See also

Examples

 fric_vatankhan(c(2118517, 2000, 2118517), c(1e-6, 70e-3/1, 7e-3/1))
#> [1] 0.01031665 0.03200000 0.03375210
 # [1] 0.01031665 0.03200000 0.03375210  # []

 fric_vatankhan(c(2118517, 5500, 2118517), c(1e-6, 50e-3/1, 7e-3/1), TRUE)
#> [1] 0.01031665 0.07556163 0.03375210
 # [1] 0.01031665 0.07556163 0.03375210