Minenergo-278. Thermal conductivity terms of pipe insulation materials

Data represent values of terms (intercept and factor) for calculating thermal conductivity of pipe insulation as a linear function of temperature of heat carrier (water). Those values are set for different insulation materials in Appendix 5.3 of Minenergo Method 278 as norms.

m278insdata

Format

A data frame with 39 rows and 4 variables:

id

Number of insulation material table 5.1 of Appendix 5.3 in Minenergo Method 278. Type: assert_integerish.

material

Designation of insulation material more or less similar to those in table 5.1 of Appendix 5.3 in Minenergo Method 278. Type: assert_character.

lambda

Value for intercept, [mW/m/°C]. Type: assert_integer.

k

Value for factor. Type: assert_integer.

Source

http://www.complexdoc.ru/ntdtext/547103/

Details

Usually the data is not used directly. Instead use function m278inshcm.

See also