Minenergo-325. Temperature drop in pipe due heat losses

Calculate temperature drop in steel pipe of district heating system (where water is a heat carrier) that is a result of heat losses through pipe wall and insulation using Minenergo Order 325 as a basis for values of heating flux.

Since Minenergo Order 325 is used as the basis for values of heating flux the calculated temperature drop may be considered as a normative temperature drop. If the actual (somehow measured) temperature drop is more than this normative temperature drop they may consider such difference to be due to extra-normative heat losses. The presence of the latter requires appropriate maintenance activities.

m325dropt(
  temperature = 130,
  pressure = mpa_kgf(6),
  consumption = 250,
  d = 700,
  len = 1,
  year = 1986,
  insulation = 0,
  laying = "underground",
  beta = FALSE,
  exp5k = TRUE
)

Arguments

temperature

temperature of heat carrier (water) inside the pipe measured at the entrance of pipe, [°C]. Type: assert_double.

pressure

absolute pressure of heat carrier (water) inside the pipe, [MPa]. Type: assert_double.

consumption

amount of heat carrier (water) that is transferred by pipe during a period, [ton/hour]. Type: assert_double.

d

internal diameter of pipe, [mm]. Type: assert_double.

len

length of pipe, [m]. Type: assert_double.

year

year when the pipe is put in operation after laying or total overhaul. Type: assert_integerish.

insulation

insulation that covers the exterior of pipe:

0

no insulation

1

foamed polyurethane or analogue

2

polymer concrete

Type: assert_subset.

laying

type of pipe laying depicting the position of pipe in space:

air,

channel,

room,

tunnel,

underground.

Type: assert_subset.

beta

should they consider additional heat losses of fittings? Type: assert_logical.

exp5k

pipe regime flag: is pipe operated more that 5000 hours per year? Type: assert_logical.

Value

normative temperature drop at the outlet of pipe, [°C]. Type: assert_double.

Details

The function is a simple wrapper for call of dropt with parameter flux calculated by m325nhl.

See also

Examples

 stopifnot(
   round(
     m325dropt(
       temperature = 130, year = 1968, laying = "channel", d = 700, l = 1000
     ), 2) == 1.37
 )