Minenergo-325. Normative heat loss data

Data represent values of specific heat loss power through the thermal insulation officially accepted by Minenergo Order 325 as norms. Those values are maximums which are legally affirmed to contribute to normative heat loss \(Q_{NHL}\) of district heating systems with water as a heat carrier.

Usage

m325nhldata

Format

A data frame with 17328 rows and 8 variables:

source

Identifier of data source: identifiers suited with glob t?p? mean appropriate table ?.? in Minenergo Order 325; identifier sgc means that values are additionally postulated (see Details). Type: assert_character.

epoch

Year depicting the epoch when the pipe is put in operation after laying or total overhaul. Type: assert_integer.

laying

Type of pipe laying depicting the position of pipe in space. Only five types of pipe laying are considered:

air,

channel,

room,

tunnel,

underground.

Type: assert_character.

exp5k

Logical indicator for pipe regime: if TRUE pipe is operated more that 5000 hours per year. Type: assert_logical.

insulation

Identifier of insulation that covers the exterior of pipe:

0

no insulation

1

foamed polyurethane or analogue

2

polymer concrete

Type: assert_integerish.

d

Nominal (outside) diameter of pipe, [mm]. Type: assert_double.

temperature

Operational temperature of pipe, [°C]. Type: assert_double.

loss

Normative value of specific heat loss power equal to heat flux output by one-meter length steel pipe during an hour, [kcal/m/h]. Type: assert_double.

Source

https://docs.cntd.ru/document/902148459

Details

Data is organized as a full factorial design, whereas for some factorial combinations Minenergo Order 325 does not provide values. For that cases values are postulated by practical reasons in Siberian cities and marked with source label sgc.

Usually the data is not used directly. Instead use function m325nhl.

See also

Examples

library(pipenostics)

head(m325nhldata)
#>   source epoch laying exp5k insulation  d temperature loss
#> 1   t1p2  1989    air FALSE          0 25          50   15
#> 2   t1p2  1989    air  TRUE          0 25          50   15
#> 3   t1p2  1989    air FALSE          0 25          75   23
#> 4   t1p2  1989    air  TRUE          0 25          75   23
#> 5   t1p2  1989    air FALSE          0 25         100   31
#> 6   t1p2  1989    air  TRUE          0 25         100   31