Diagnostics, reliability and predictive maintenance of pipeline systems

The motivation for the package was to aggregate to some extent the separate knowledge about engineering, reliability, diagnostics and predictive maintenance of pipeline systems suitable for pipe holders. Distributing such knowledge with R-package seemed the most attractive option for white-collar engineers, having utilized spreadsheet software as a mainstream.

Details

Aiming to avoid portability and accessibility problems made us search ways to restrict source code development by functionality of few external packages.

Since most functions have native argument vectorization usage of those functions with fast data.table framework is strongly encourage when processing large data sets.

For that purpose arguments for all package functions are thoroughly checked for type consistency and physical sense using asserts and tests from checkmate package. Moreover, in package documentation we borrow type designations according to checkmate notation.

1. Corrosion diagnostics

The next values describing technological conditions, material properties of pipe and defect parameters are used as arguments throughout the most functions concerning corrosion diagnostics:

maop

maximum allowable operating pressure - MAOP, [PSI]

d

nominal outside diameter of the pipe, [inch]

wth

nominal wall thickness of the pipe, [inch]

smys

specified minimum yield of stress (SMYS) as a characteristics of steel strength, [PSI]

depth

measured maximum depth of the corroded area, [inch]

l

measured maximum longitudial length of the corroded area, [inch]

1.1 ASME B31G-1991

It is recognized by pipeline companies that some sections of high pressure pipelines particularly those installed a number of years ago, have experienced some corrosion. Where corrosion is found, pipeline operators have been deeply concerned about the need for a method of determining the remaining strength of these corroded areas. If the corrosion does not penetrate the pipe wall, what is the pressure containing capability of the remaining pipe metal in terms of its ability to continue to operate safely at the maximum allowable operating pressure (MAOP) of the pipeline system?

Thus, one of the needs of the pipeline industry has been a procedure that will help operators, particularly field personnel, make decisions on existing pipelines, when exposed for any purpose, as to whether any corroded region may be left in service or whether it needs to be repaired or replaced. Such determinations must be based upon sound research and extensive testing in order to provide safe and conservative guidelines on which to base field decisions.

The ASME B31G-1991 provides procedures to assist in this determination.

Appendix A to ASME B31G-1991 shows the source code for determining the allowable length and maximum allowable working pressure. The b31g* and b31crvl functions reproduce the idea of CRVL.BAS. They are natively vectorized.

Usage of b31crvl function that imitates the output of _CRVL.BAS_ is presentaed in Example 1.

1.2 ASME B31G-2012

An effort was undertaken to update the ASME B31G-1991 up to ASME B31G-2012 document to recognize certain other corrosion evaluation methods that have proven sound and that have seen successful use in the pipeline industry. Incorporation of these other methods provides us with a formalized framework within which to use such methodologies.

Nevertheless, to preserve simplicity of traditional inline measurements during inspections we only consider Analysis Level 1.

As noted in ASME B31G-2012, Level 1 evaluation is quite suitable for use in prioritizing corrosion defects identified by inline inspection.

1.3 Other models

Other approaches for operating with corrosion data are mostly aimed on failure pressure calculations. Models like dnvpf, shell92pf, and pcorrcpf assume different shapes of corrosion defects and usage conditions for some cases. So, it is encouraged first to find out which model is most suitable for solving some real world problem.

2. Heat losses

Heat loss is the energy characteristic of district heating networks. It is the amount of heat energy spent on the transportation and distribution of heat energy from the source to the consumers.

Heat losses depend on the operating temperature, technical condition, volume and configuration of the district heating network, as well as on climatic factors. Heat losses are additive being the sum of the heat losses of individual pipeline segments.

Determination of heat losses for pipeline segments hereinafter is called heat loss localization.

It is assumed that actual heat loss (AHL) of pipeline segment has two contributions: normative heat loss (NHL) and extra-normative heat loss (ExNHL).So we can write:

$$ AHL = NHL + ExNHL, ExNHL > 0. $$

Localization of ExNHL is an important part of health maintenance activities of district heating network operation. One can determine ExNHL of pipeline segment as a positive difference between AHL and NHL and it is the most natural way. For that purpose Minenergo-325 and Minenergo-278 methods for postulating NHL are considered. Minenergo-325 lists legally affirmed maximum values of heat flux that is allowed to be emitted by steel pipes (see m325nhl). Higher emission is treated as ExNHL. Minenergo-278 gives method for engineering calculation of NHL considering technical condition of pipeline segment (see m278hlcha, m278hlund, and m278hlair).

Examples

## Example 1
b31crvl(maop = 910, d = 30, wth = .438, smys = 52000, def  = .72, depth = .1, l = 7.5)

#> 
#> -- Calculated data --
#> Intermediate factor (A) = 1.847
#> Design pressure = 1093 PSI; Safe pressure = 1093 PSI
#> Pipe may be operated safely at MAOP, 910 PSI
#> With corrosion length 7.500 inch, maximum allowed corrosion depth is 0.2490 inch; A = 1.847
#> With corrosion depth 0.100 inch, maximum allowed corrosion length is Inf inch; A = 5.000

 # -- Calculated data --
 # Intermediate factor (A) = 1.847
 # Design pressure = 1093 PSI; Safe pressure = 1093 PSI
 # Pipe may be operated safely at MAOP, 910 PSI
 # With corrosion length 7.500 inch, maximum allowed corrosion depth is 0.2490 inch; A = 1.847
 # With corrosion depth 0.100 inch, maximum allowed corrosion length is Inf inch; A = 5.000