Estimate Reynolds number

Estimate Reynolds number for fluid flow in a cylindrical pipe.

Usage

re_u(d, mu, u, rho)

re_v(d, mu, v, rho)

re_m(d, mu, m)

Arguments

d

internal diameter of pipe, [m]. Type: assert_double.

mu

dynamic viscosity of fluid in pipe, [kg/m/s]. Type: assert_double.

u

mean velocity of fluid in pipe, [m/s]. Type: assert_double.

rho

mass density of fluid in pipe, [kg/m³]. Type: assert_double.

v

volumetric flow rate of fluid in pipe, [m³/s]. Type: assert_double.

m

mass flow rate of fluid in pipe, [kg/s]. Type: assert_double.

Value

Reynolds number - a dimensionless quantity that reveals the ratio between inertial and viscous forces in the fluid, []. Type: assert_double.

Details

The calculation of Reynolds number is bounded by physically reasonable limits of fluid properties found in domain specificity of the package.

See also

Examples

library(pipenostics)

# Reynolds numbers for typical district heating water flows at temperature
# near 25 °C in a set of pipes with different sizes:

range(re_u(seq(.25, 1, 0.05), .89, 1, 1000))
#> [1]  280.8989 1123.5955