Shell92. Failure pressure of the corroded pipe

Calculate failure pressure of the corroded pipe according to Shell92 code.

This code should be applied only to

single cross section of the pipeline containing a longitudinally oriented, flat bottom surface defect of the corrosion/erosion type;

defects which depth is less than 85 % of pipe wall thickness.

The estimation is valid for single isolated metal loss defects of the corrosion/erosion type and when only internal pressure loading is considered.

As in the case of dnvpf, the defect is approximated by a rectangular form.

shell92pf(d, wth, uts, depth, l)

Arguments

d

nominal outside diameter of the pipe, [mm]. Type: assert_double.

wth

nominal wall thickness of the pipe, [mm]. Type: assert_double.

uts

ultimate tensile strength (UTS) or specified minimum tensile strength (SMTS) as a characteristic of steel strength, [MPa]. Type: assert_double.

depth

measured maximum depth of the corroded area, [mm]. Type: assert_double.

l

measured maximum longitudinal length of corroded area, [mm]. Type: assert_double.

Value

Estimated failure pressure of the corroded pipe, [MPa]. Type: assert_double.

Details

Numeric NAs may appear in case prescribed conditions of use are offended.

References

Timashev and A. Bushinskaya, Diagnostics and Reliability of Pipeline Systems, Topics in Safety, Risk, Reliability and Quality 30, DOI 10.1007/978-3-319-25307-7

See also

Examples

d     = c(812.8, 219.0)  # [mm]
wth   = c( 19.1,  14.5)  # [mm]
uts  = c(530.9, 455.1)   # [N/mm^2]
l     = c(203.2, 200.0)  # [mm]
depth = c( 13.4,   9.0)  # [mm]

shell92pf(d, wth, uts, depth, l)
#> [1] 11.09262 25.27286
# [1] 11.09262 25.27286