ASME B31G. Corrosion state of 12 pipes

Data represents examples used for verification of computer program CRVL.BAS listed in Appendix A of ASME B31G-1991.

b31gdata

Format

A data frame with 12 rows and 15 variables:

maop

maximum allowable operating pressure - MAOP, [PSI]. Type: assert_double.

d

nominal outside diameter of the pipe, [inch]. Type: assert_double.

wth

nominal wall thickness of the pipe, [inch]. Type: assert_double.

smys

specified minimum yield of stress (SMYS) as a characteristics of steel strength, [PSI]. Type: assert_double.

def

appropriate (combined) design factor from ASME B31.4, ASME B31.8, or ASME B31.11, []. Type: assert_double.

depth

measured maximum depth of the corroded area, [inch]. Type: assert_double.

l

measured maximum longitudinal length of corroded area, [inch]. Type: assert_double.

status

Operational status of pipe: 1 - excellent, 2 - monitoring is recommended, 3 - alert! replace the pipe immediately! Type: assert_numeric.

design_pressure

design pressure of the pipe, [PSI]. Type: assert_double.

safe_pressure

safe maximum pressure for the corroded area, [PSI]. Type: assert_double.

pressure_exceeding

whether operator's action is required to reduce MOAP lower than the maximum safe pressure of the corroded area. . Type: assert_logical.

allowed_corrosion_depth

allowable depth of the corroded area, [inch]. Type: assert_double.

A

intermediate factor related to the geometry of the corroded area, []. Type: assert_double.

allowed_corrosion_length

allowable length of the corroded area, [inch]. Type: assert_double.

AP

another intermediate factor related to the geometry of the corroded area, []. Type: assert_double.

Source

https://law.resource.org/pub/us/cfr/ibr/002/asme.b31g.1991.pdf