Temperature drop in cylindrical steel pipe due heat loss

Calculate temperature drop in steel pipe of district heating system (where water is a heat carrier) that is a result of heat loss through pipe wall and insulation.

Usage

dropt(
  temperature = 130,
  pressure = mpa_kgf(6),
  flow_rate = 250,
  loss_power = 7000
)

Arguments

temperature

temperature of heat carrier (water) inside the pipe measured at the inlet of pipe, [°C]. Type: assert_double.

pressure

absolute pressure of heat carrier (water) inside the pipe, [MPa]. Type: assert_double.

flow_rate

amount of heat carrier (water) that is transferred by pipe during a period, [ton/h]. Type: assert_double.

loss_power

power of heat loss - heat loss through area of pipe wall per hour, [kcal/h]. Type: assert_double.

Value

Temperature drop at the outlet of pipe, [°C]. Type: assert_double.

Details

Specific isobaric heat capacity used in calculations is calculated according to IAPWS R7-97(2012) for Region 1 since it is assumed that state of water in district heating system is always in that region.

See also

Examples

library(pipenostics)

# Calculate normative temperature drop based on Minenergo-325 for pipe
# segment
pipeline <- list(
  year   = 1968,
  laying = "channel",
  d      = 700, # [mm]
  len    = 1000 # [m]
)

regime <- list(
  temperature = c(130, 150), # [°C]
  pressure    = .588399,     # [MPa]
  flow_rate   = 250          # [ton/h]
)

pipe_loss_power <- do.call(
    m325nhl,
    c(pipeline, temperature = list(regime[["temperature"]]), duration = 1)
)

temperature_drop <- dropt(
  temperature = regime[["temperature"]], # [°C]
  loss_power  = pipe_loss_power          # [kcal/h]
)                                        # [°C]

print(temperature_drop)
#> [1] 1.366806 1.433840