Temperature drop in pipe due heat losses

Calculate temperature drop in steel pipe of district heating system (where water is a heat carrier) that is a result of heat losses through pipe wall and insulation.

dropt(temperature = 130, pressure = mpa_kgf(6), consumption = 250, flux = 7000)

Arguments

temperature
temperature of heat carrier (water) inside the pipe measured at the inlet of pipe, [°C]. Type: assert_double.

pressure
absolute pressure of heat carrier (water) inside the pipe, [MPa]. Type: assert_double.

consumption
amount of heat carrier (water) that is transferred by pipe during a period, [ton/hour]. Type: assert_double.

flux
heat flux emitted by pipe during a period, [kcal/hour]. Type: assert_double.

Value

temperature drop at the outlet of pipe, [°C]. Type: assert_double.

Details

Specific isobaric heat capacity used in calculations is calculated according to IAPWS R7-97(2012) for Region 1 since it is assumed that state of water in district heating system is always in that region.

See also

Examples

 # Calculate normative temperature drop based on Minenergo-325 for pipe segment
 pipeline <- list(
   year = 1968,
   laying = "channel",
   d = 700,
   l = 1000
 )
 operation_temperature <- c(130, 150)  # [°C]

 foo <- dropt(
   temperature = operation_temperature,
   flux = do.call(
     m325nhl,
     c(pipeline, temperature = list(operation_temperature))
   )
 )

 foo

#> [1] 1.366806 1.433840

 # 1.37 [°C]

 # This is the same as using m325dropt:
 bar <- m325dropt(temperature = operation_temperature,
   year = 1968, laying = "channel", d = 700, l = 1000
 )

 stopifnot(all.equal(foo, bar))