Minenergo-278. Normative heat loss of open-air pipe

Calculate normative heat loss through the thermal insulation of the open-air supplying pipe as a function of construction, operation, and technical condition specifications according to Appendix 5.1 of Minenergo Method 278.

This type of calculations is usually made on design stage of district heating network (where water is a heat carrier) and is closely related to building codes and regulations.

Usage

m278hlair(
  t1 = 110,
  t2 = 60,
  t0 = 5,
  insd1 = 0.1,
  insd2 = insd1,
  d1 = 0.25,
  d2 = d1,
  lambda1 = 0.09,
  lambda2 = 0.07,
  k1 = 1,
  k2 = k1,
  lambda0 = 26,
  len = 1,
  duration = 1
)

Arguments

t1

temperature of heat carrier (water) inside the supplying pipe, [°C]. Type: assert_double.

t2

temperature of heat carrier (water) inside the returning pipe, [°C]. Type: assert_double.

t0

temperature of environment, [°C]. In case of open-air pipe this is the ambient temperature. Type: assert_double.

insd1

thickness of the insulator which covers the supplying pipe, [m]. Type: assert_double.

insd2

thickness of the insulator which covers the returning pipe, [m]. Type: assert_double.

d1

outside diameter of supplying pipe, [m]. Type: assert_double.

d2

outside diameter of returning pipe, [m]. Type: assert_double.

lambda1

thermal conductivity of insulator which covers the supplying pipe [W/m/°C]. Type: assert_double.

lambda2

thermal conductivity of insulator which covers the returning pipe [W/m/°C]. assert_double.

k1

technical condition factor for insulator of supplying pipe, []. Type: assert_double.

k2

technical condition factor for insulator of returning pipe, []. Type: assert_double.

lambda0

thermal conductivity of environment, [W/m/°C]. In case of overhead laying this is the thermal conductivity of open air. Type: assert_double.

len

length of supplying pipe, [m]. Type: assert_double.

duration

duration of heat loss, [h]. Type: assert_double.

Value

Normative heat loss of the open-air laid supplying cylindrical pipe during duration, [kcal]. If len of pipe is 1 m (meter) as well as duration is set to 1 h (hour) (default values) then the return value is also the specific heat loss power, [kcal/m/h] and so comparable with those prescribed by Minenergo Order 325. Type: assert_double.

Details

Details on using k1 and k2 are the same as for m278hlcha.

See also

Examples

library(pipenostics)

m278hlair()
#> [1] 138.7736