Minenergo-278. Thermal conductivity of pipe insulation materials

Get normative values of thermal conductivity of pipe insulation materials affirmed by Minenergo Method 278 as a function of temperature of heat carrier (water).

Usage

m278inshcm(temperature = 110, material = "aerocrete")

Arguments

temperature

temperature of heat carrier (water) inside the pipe, [°C]. Type: assert_double.

material

designation of insulation material as it stated in m278insdata. Type: assert_subset.

Value

Thermal conductivity of insulation materials at given set of temperatures, [W/m/°C], [W/m/K]. Type: assert_double.

See also

Examples

library(pipenostics)

# Range of thermal conductivity of pipe insulation at 110 °C:

range(m278inshcm(110, m278insdata[["material"]]))
#> [1] 0.05000 0.15725