Minenergo-278. Thermal conductivity of pipe insulation materials

Get normative values of thermal conductivity of pipe insulation materials affirmed by Minenergo Method 278 as a function of temperature of heat carrier (water).

m278inshcm(temperature = 110, material = "aerocrete")

Arguments

temperature
temperature of heat carrier (water) inside the pipe, [°C]. Type: assert_double.

material
designation of insulation material as it stated in m278insdata, Type: assert_subset.

Value

Thermal conductivity of insulation materials [W/m/°C] at given set of temperatures. Type: assert_double.

See also

Examples

# Averaged thermal conductivity of pipe insulation at 110 °C
print(m278insdata)

#>    id                 material lambda   k
#> 1   1        asbestic_sovelite     87 120
#> 2   2      asbestic_glassfiber     58 230
#> 3   3           asbestic_cloth    130 260
#> 4   4            asbestic_cord    120 310
#> 5   5     asbestic_cord_common    130 260
#> 6   6       asbestic_cord_lint     93 200
#> 7   7 asbestic_vermiculite_250     81 200
#> 8   8 asbestic_vermiculite_300     87 230
#> 9   9      bituminous_pearlite    120 230
#> 10 10          bituminous_clay    130 230
#> 11 11   bituminous_vermiculite    130 230
#> 12 12            volcanite_300     74 150
#> 13 13            diatomite_500    116 230
#> 14 14            diatomite_600    140 230
#> 15 15       calcic_silicon_200     69 150
#> 16 16         mineral_wool_100     45 200
#> 17 17         mineral_wool_125     49 200
#> 18 18         mineral_wool_075     43 220
#> 19 19             glasfiber_00     40 260
#> 20 20             glasfiber_50     42 280
#> 21 21                aerocrete    110 300
#> 22 22                  plastic     43 190
#> 23 23         polymer_concrete     70   0
#> 24 24             polyurethane     50   0
#> 25 25       perlite_cement_300     76 185
#> 26 26       perlite_cement_350     81 185
#> 27 27   mineral_wool_100_block     44 210
#> 28 28   mineral_wool_125_block     47 185
#> 29 29         mineral_wool_250     56 185
#> 30 30             glasfiber_75     44 230
#> 31 31         mineral_wool_150     49 200
#> 32 32         mineral_wool_200     52 185
#> 33 33             sovelite_350     76 185
#> 34 34             sovelite_400     78 185
#> 35 35             mineral_wool     69 190
#> 36 36           porous_plastic     50   0
#> 37 37    mineral_wool_200_cord     56 185
#> 38 38    mineral_wool_250_cord     58 185
#> 39 39    mineral_wool_300_cord     61 185

stopifnot(
  round(
    mean(with(m278insdata, { m278inshcm(110, material)})),
    2) == 9e-2
)

# Terms for linear connection between thermal conductivity of unknown
# (averaged) pipe insulator vs temperature:
temperature <- as.double(1:450)
lambda_ins <- with(m278insdata, {
  vapply(temperature, function(x) mean(m278inshcm(x, material)), .1)
})
C <- coef(lsfit(temperature, lambda_ins))  # c(Intercept, X)
stopifnot(
  all(abs(C - c(7.963590e-02, 9.730769e-05)) < 1e-8)
)