Minenergo-325. Local heat loss coefficient

Calculate \(\beta\) - local heat loss coefficient according to rule 11.3.3 of Minenergo Order 325. Local heat loss coefficient is used to increase normative heat losses of pipe by taking into account heat losses of fittings (shut-off valves, compensators and supports). This coefficient is applied mostly as a factor during the summation of heat losses of pipes in pipeline leveraging formula 14 of Minenergo Order 325.

m325beta(laying = "channel", d = 700)

Arguments

laying

type of pipe laying depicting the position of pipe in space:

air,

channel,

room,

tunnel,

underground.

Type: assert_subset.

d

internal diameter of pipe, [mm]. Type: assert_double.

Value

Two possible values of \(\beta\): 1.2 or 1.15 depending on pipe laying and its diameter. Type: assert_double.

See also

Examples

norms <- within(m325nhldata, {
  beta <- m325beta(laying, as.double(diameter))
})
unique(norms$beta)
#> [1] 1.15 1.20
# [1] 1.15 1.20