Minenergo-325. Data for normative heat losses of pipe

Data represent values of heat losses officially accepted by Minenergo Order 325 as norms. Those values represent heat flux that is legally affirmed to be emitted per meter during an hour by steel pipe of district heating system with water as a heat carrier.

m325nhldata

Format

A data frame with 17328 rows and 8 variables:

source

Identifier of data source: identifiers suited with glob t?p? mean appropriate table ?.? in Minenergo Order 325; identifier sgc means that values are additionally postulated (see Details). Type: assert_character.

epoch

Year depicting the epoch when the pipe is put in operation after laying or total overhaul. Type: assert_integer.

laying

Type of pipe laying depicting the position of pipe in space. Only five types of pipe laying are considered:

air,

channel,

room,

tunnel,

underground.

Type: assert_character.

exp5k

Logical indicator for pipe regime: if TRUE pipe is operated more that 5000 hours per year. Type: assert_logical.

insulation

Identifier of insulation that covers the exterior of pipe:

0

no insulation

1

foamed polyurethane or analogue

2

polymer concrete

Type: assert_integerish.

diameter

Nominal internal diameter of pipe, [mm]. Type: assert_double.

temperature

Operational temperature of pipe, [°C]. Type: assert_double.

flux

Heat flux emitted by every meter of pipe during an hour, [kcal/m/hour]. Type: assert_double.

Source

http://docs.cntd.ru/document/902148459

Details

Data is organized as a full factorial design, whereas for some factorial combinations Minenergo Order 325 does not provide values. For that cases values are postulated by practical reasons in Siberian cities and marked with source label sgc.

Usually the data is not used directly. Instead use function m325nhl.

See also