Get list of weather stations (meteos)

Get a list of weather stations located primarily in the central and northern parts of Eurasia. For each weather station, the following information is provided: an integer station ID, geographic coordinates, altitude, and the mean annual ground temperature averaged over depth.

Usage

meteos()

Value

List of weather stations (meteos) with the next fields:

station_id

Weather station unique identifier. Type: assert_integer.

name

Human-readable name of weather station. Type: assert_character.

lat

Geographical position of weather station. Latitude, [DD]. Type: assert_double.

lon

Geographical position of weather station. Longitude, [DD]. Type: assert_double.

alt

Altitude - position of weather station above sea level, [m]. Type: assert_double.

avg

Mean annual ground temperature averaged over depth, [°C]. Type: assert_double.

Type: assert_data_frame.

References

Climate Change Investigation Laboratory. Description of the array of daily data on soil temperature at depths up to 320 cm by meteorological stations of the Russian Federation.

See also

Examples

library(pipenostics)

head(meteos())
#>       station_id           name   lat   lon alt      avg
#> 22217      22217    Kandalaksha 67.15 32.35  26 4.299446
#> 22408      22408       Kalevala 65.22 31.15 118 4.674742
#> 22471      22471          Mezen 65.87 44.22  14 3.544074
#> 22583      22583         Koinas 64.75 47.65  63 4.817047
#> 22981      22981 Velikiy-Ustyug 60.77 46.30  93 5.975453
#> 23405      23405     Ust-Tsilma 65.43 52.27  78 4.094936