PCORRC. Failure pressure of the corroded pipe

Calculate failure pressure of the corroded pipe according to PCORRC model.

PCORRC methodology was developed on the basis of studying the mechanism of destruction of pipes, material of which has improved or high fracture toughness, and on the high-precision modeling of the finite element pipe models performed at the Battelle Institute. According to field test results of a large number of actual pipe segments, the destruction mechanism for defective pipeline segment depends on the pipe material fracture toughness. These tests also showed that only pipes made out of steel with improved or high fracture toughness fail a result of plastic fracture. In determining the Folias factor the effect of increased stress concentration and steel hardening in the plastic deformation zone at the start of the defect failure process was taken into account.

This code should be applied only to

a single cross section of the pipeline containing a longitudinally oriented, flat bottom surface defect of the corrosion/erosion type;

pipelines, which operate at temperatures exceeding the temperature of pipe material ductile–brittle transition, and for pipematerial with the impact energy of Charpy 61 [J] and above.

pcorrcpf(d, wth, uts, depth, l)

Arguments

d
nominal outside diameter of the pipe, [mm]. Type: assert_double.

wth
nominal wall thickness of the pipe, [mm]. Type: assert_double.

uts
ultimate tensile strength (UTS) or specified minimum tensile strength (SMTS) as a characteristic of steel strength, [\(MPa\)]. Type: assert_double.

depth
measured maximum depth of the corroded area, [mm]. Type: assert_double.

l
measured maximum longitudinal length of corroded area, [mm]. Type: assert_double.

Value

Estimated failure pressure of the corroded pipe, [\(MPa\)]. Type: assert_double.

References

S. Timashev and A. Bushinskaya, Diagnostics and Reliability of Pipeline Systems, Topics in Safety, Risk, Reliability and Quality 30, DOI 10.1007/978-3-319-25307-7

A.C.Reddy, Safety Failure Criteria of Fluorocarbon Plastic Pipes for Dry Chlorine Transport using Finite Element Analysis Materials today: proceedings, Vol. 4(8), 2017, pp. 7498-7506.

See also

Examples

 ## unit test:
 with(data.frame(
   d     = c(812.8, 219.0),  # [mm]
   wth   = c( 19.1,  14.5),  # [mm]
   uts  = c(530.9, 455.1),  # [N/mm^2]
   l     = c(203.2, 200.0),  # [mm]
   depth = c( 13.4,   9.0)   # [mm]
   ),
     stopifnot(
        all(round(pcorrcpf(d, wth, uts, depth, l), 4) == c(16.3545, 33.0129))
     )
 )