DNV-RP-F101. De-rate yield stress and tensile strength of pipe due to temperature

Temperature is highly influence on pipe material properties and especially on its strength. Since in API SPECIFICATION 5L values of SMYS or UTS are postulated at room conditions, in case of higher temperature magnitudes they should be corrected. For that purpose DNV-RP-F101 offers linear de-rating for SMYS or SMYS according to figure 2-3.

strderate(x, temperature = 24.3)

Arguments

x

specified minimum yield of stress (SMYS), or ultimate tensile strength (UTS), or specified minimum tensile strength (SMTS) as a characteristic of steel strength at room temperature, [MPa]. Type: assert_double.

temperature

temperature of pipe wall, [°C]. Type: assert_double.

Value

de-rated value of x, i.e. of appropriate pipe material property, [MPa] . Type: assert_double.

See also

Examples

with(api5l3t, {
print(strderate(mpa_psi(smys), 53))
print(
  strderate(mpa_psi(uts),seq(0, 250, length.out = length(smys)))
)
})
#>  [1] 170.5689 205.0427 239.5165 287.7798 315.3588 356.7274 384.3064 411.8854
#>  [9] 446.3592 480.8330 549.7806
#>  [1] 310.2641 330.9483 413.6854 398.6854 404.3697 415.0540 439.5278 457.1068
#>  [9] 460.8963 485.3701 530.5282
# [1] 170.5689 205.0427 239.5165 287.7798 315.3588 356.7274 384.3064 411.8854 446.3592 480.8330
# [11] 549.7806
# [1] 310.2641 330.9483 413.6854 398.6854 404.3697 415.0540 439.5278 457.1068 460.8963 485.3701
# [11] 530.5282